added structure + CSS
