+++ src/index.css
@@ -11,3 +11,252 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
 }
+
+:root {
+  --general-padding: 35px;
+  --general-smaller-padding: 20px;
+  --general-mini-padding: 5px;
+  --box-shadow: 8px 8px 0px 0px rgba(0, 0, 0, 1);
+  --main-width-mobile: 90%;
+  --main-width-desktop: 70%;
+  --hover-color: rgb(255, 199, 220)
+}
+
+.emoji {
+  font-size: 24px
+}
+
+p {
+  word-break: break-all;
+  white-space: normal
+}
+
+a {
+  text-decoration: none;
+  color: #000
+}
+
+.main-wrapper {
+  margin: 20px auto;
+  padding: 5px var(--general-padding);
+  width: 80%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 25px
+}
+
+header {
+  margin: 0;
+  padding: 0 var(--general-padding);
+  text-align: center
+}
+
+h1 {
+  font-family: Abril Fatface;
+  font-size: 1.5em;
+  font-weight: 800;
+  background: -webkit-linear-gradient(45deg, #ff199f, #cd8cdd, #5be288, #bef146, #7e6ed8);
+  background-clip: text;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  margin: 0
+}
+
+.post-wrapper {
+  background-color: #ebebeb;
+  border: 1px solid rgb(105, 105, 105);
+  margin: 0;
+  padding: 5px var(--general-padding) 15px var(--general-padding);
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-start;
+  gap: 10px;
+  box-shadow: var(--box-shadow);
+  -webkit-box-shadow: var(--box-shadow);
+  width: var(--main-width-mobile)
+}
+
+.post-wrapper form {
+  width: 100%
+}
+
+.post-wrapper h2 {
+  font-size: 26px;
+  margin: 5px 0 0
+}
+
+.post-wrapper textarea {
+  font-size: 18px;
+  width: 100%;
+  margin: 10px 0 2px;
+  background: white;
+  text-align: center
+}
+
+.post-length {
+  display: flex;
+  justify-content: space-between;
+  margin: 0 0 20px;
+  padding: 0;
+  gap: 25px
+}
+
+.length {
+  color: gray;
+  margin: 0;
+  min-width: 50px
+}
+
+.error {
+  font-size: 16px;
+  color: gray;
+  margin: 0
+}
+
+.post-wrapper button {
+  background-color: #ffb7d2;
+  border-radius: 25px;
+  padding: 5px 10px;
+  margin: 0 0 5px;
+  border: none;
+  font-size: 18px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 5px;
+  cursor: pointer
+}
+
+.post-wrapper button:hover {
+  background-color: var(--hover-color)
+}
+
+.list-wrapper {
+  margin: 0;
+  padding: 5px var(--general-padding);
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 25px
+}
+
+.message {
+  border: 1px solid rgb(105, 105, 105);
+  background-color: #fff;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  padding: var(--general-smaller-padding);
+  box-shadow: var(--box-shadow);
+  -webkit-box-shadow: var(--box-shadow)
+}
+
+.message p {
+  margin: 10px 0 15px;
+  font-size: 18px
+}
+
+.info-wrapper {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 5px
+}
+
+.like-button {
+  background-color: #d1d1d1;
+  padding: 10px;
+  border-radius: 50%;
+  border: none;
+  cursor: pointer;
+  animation: pulse-small 2s infinite
+}
+
+.like-button:hover {
+  background-color: var(--hover-color)
+}
+
+.num-likes {
+  margin-left: 5px
+}
+
+.info-time {
+  color: gray
+}
+
+footer {
+  margin: 0;
+  padding: 5px var(--general-padding);
+  display: flex;
+  flex-direction: column;
+  align-items: center
+}
+
+footer .emoji {
+  font-size: 42px;
+  animation: pulse-big 2s infinite;
+  margin-bottom: 10px
+}
+
+a:hover {
+  font-weight: 700
+}
+
+.liked {
+  background-color: #ee8bb4
+}
+
+.red {
+  color: red
+}
+
+@keyframes pulse-small {
+  0% {
+    transform: scale(1)
+  }
+
+  50% {
+    transform: scale(1.1)
+  }
+
+  to {
+    transform: scale(1)
+  }
+}
+
+@keyframes pulse-big {
+  0% {
+    transform: scale(1)
+  }
+
+  50% {
+    transform: scale(1.5)
+  }
+
+  to {
+    transform: scale(1)
+  }
+}
+
+@media (min-width: 675px) {
+  .main-wrapper {
+    width: 60%
+  }
+
+  .post-wrapper,
+  .list-wrapper {
+    width: var(--main-width-desktop)
+  }
+
+  .message {
+    width: 100%
+  }
+
+  h1 {
+    font-size: 45px
+  }
+}
